Separate PhalyfusionGlobalInspection from PhalyfusionValidationInspection
PhalyfusionGlobalInspection is used for batch inspections, "Analyse code" action or "Run inspection by name" action (or just a button in toolbar). It launches Phalyfusion for all files in one run (in multiple runs at Windows).

PhalyfusionValidationInspection is used to highlight cached problems from PhalyfusionGlobalInspection on-fly. Yet there is a special setting "Launch Phalyfusion in on-fly mode" to make it run Phalyfusion tool by itself on-fly